+++ packages/sdk/src/lib/CasperHDKey.ts
@@ -51,7 +51,10 @@ export class CasperHDKey {
    */
   public derive(path: string): Secp256K1 {
     const secpKeyPair = this.hdKey.derive(path);
-    return new Secp256K1(secpKeyPair.publicKey!, secpKeyPair.privateKey!);
+    return new Secp256K1(
+      new Uint8Array(secpKeyPair.publicKey!), 
+      new Uint8Array(secpKeyPair.privateKey!)
+    );
   }
 
   /**