+++ packages/sdk/CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to casper-client-sdk.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.0.21
+
+### Added
+
+- `CasperClient.getDeployByHashFromRPC` allows for getting `Deploy` instance from the Node's RPC.
+
+### Fixed
+
+- Secp keys generator returns `Uint8Array` instead of `Buffer`.
+
+### Changed
+
+- `CLValue.publicKey` accepts `PublicKey` object.
+
 ## 1.0.20
 
 ### Fixed